Utility\InputValidatorTest: split off `is_iterable()` tests to dedicated test class
What remains are the tests for the `is_curl_handle()` method, so renaming the original test class to reflect this.
